api/requirements.txt
flask==3.1.3
flask-cors==5.0.1
mysql-connector-python==9.1.0
redis==5.2.1
pyyaml==6.0.2
requests==2.33.0
apscheduler==3.10.4
